# docs/conhecimento/anexos-openapi-mvp.yaml
openapi: 3.1.0
info:
  title: SGS MVP API
  version: 0.1.0-mvp
  description: |
    Contratos preliminares do Sistema de Gestão em Saúde (Franca/SP) — MVP.
    Tecnologicamente neutros. Alinhados às specs em data/esus/5.5.24/spec/.
    NÃO implementar UI aqui; design ainda pendente.
  contact:
    name: SIGS
servers:
  - url: /api/v1
    description: API local/homologação
...
paths:
...
  /vaccinations:
    post:
      tags: [vaccinations]
      summary: Registrar vacinação
      operationId: createVaccination
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VaccinationWrite'
      responses:
        '201':
          description: Criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VaccinationRecord'
        '400':
          $ref: '#/components/responses/ValidationError'
  /vaccinations/{vaccinationId}:
    get:
      tags: [vaccinations]
      summary: Obter registro de vacinação
      operationId: getVaccination
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/vaccinationId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VaccinationRecord'
  /patients/{patientId}/vaccination-card:
    get:
      tags: [vaccinations]
      summary: Cartão de vacina do paciente
      operationId: getVaccinationCard
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/patientId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VaccinationCard'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    patientId:
      name: patientId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    vaccinationId:
      name: vaccinationId
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
...
    ValidationError:
      description: Erro de validação de domínio
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        fields:
          type: object
          additionalProperties: { type: string }
...
    VaccinationStrategy:
      type: string
      description: Códigos alinhados a EstrategiaVacinacaoDbEnum / LEDI
      enum:
        - ROUTINE
        - SPECIAL
        - BLOCK
        - INTENSIFICATION
        - CAMPAIGN_INDISCRIMINATE
        - CAMPAIGN_SELECTIVE
        - SEROTHERAPY
        - PRIVATE
        - MEV
        - MULTIVACCINATION
        - RESEARCH
        - PRE_EXPOSURE
        - POST_EXPOSURE
        - REEXPOSURE
        - SCHOOL
    VaccineApplication:
      type: object
      required: [immunobiologicalId, attendanceGroupId, lot, manufacturerId, routeId, siteId]
      properties:
        immunobiologicalId: { type: string }
        strategy:
          $ref: '#/components/schemas/VaccinationStrategy'
        attendanceGroupId: { type: string }
        doseId: { type: string }
        lot: { type: string, maxLength: 30 }
        manufacturerId: { type: string }
        routeId: { type: string }
        siteId: { type: string }
        prescriberCboId: { type: string }
        indicationCid10: { type: string }
        leprosyContact: { type: boolean }
        isClinicalResearch: { type: boolean }
        anvisaStudyProtocol: { type: string, maxLength: 10 }
        anvisaProtocolVersion: { type: string, maxLength: 10 }
        anvisaRegistrationNumber: { type: string, maxLength: 10 }
    VaccinationWrite:
      type: object
      required: [patientId, applications]
      properties:
        patientId: { type: string, format: uuid }
        shift: { type: string }
        careLocationId: { type: string }
        applications:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/VaccineApplication'
    VaccinationRecord:
      type: object
      properties:
        id: { type: string, format: uuid }
        patientId: { type: string, format: uuid }
        status:
          type: string
          enum: [DRAFT, NOT_SENT, SENT]
        applications:
          type: array
          items:
            $ref: '#/components/schemas/VaccineApplication'
        createdAt: { type: string, format: date-time }
    VaccinationCard:
      type: object
      properties:
        patientId: { type: string, format: uuid }
        patientName: { type: string }
        doses:
          type: array
          items:
            type: object
            properties:
              date: { type: string, format: date }
              immunobiological: { type: string }
              dose: { type: string }
              lot: { type: string }
              strategy: { type: string }
...